Request all blog access after reloading proxy
